refactor: convert userRouter to ts
